refactor psrctl files
